+++ packages/fighting-design/skeleton/src/props.ts
@@ -1,4 +1,4 @@
-import { setBooleanProp, setNumberProp, setStringProp } from '../../_utils'
+import { setBooleanProp, setNumberProp, setStringNumberProp, setStringProp } from '../../_utils'
 import { FIGHTING_SIZE } from '../../_tokens'
 import type { ExtractPropTypes } from 'vue'
 import type { FightingSize } from '../../_interface'
@@ -6,12 +6,16 @@ import type { FightingSize } from '../../_interface'
 export const Props = {
   /** 是否带圆角的 */
   round: setBooleanProp(),
+  /** 是否带有差异长度 */
+  difference: setBooleanProp(),
   /** 是否展示波浪动画 */
   animated: setBooleanProp(),
   /** 渲染多行占位图 */
   rows: setNumberProp(1),
   /** 为 true 时，显示占位图。false 展示子组件 */
   loading: setBooleanProp(),
+  /** 间距 */
+  rowGap: setStringNumberProp(),
   /**
    * 尺寸
    *