+++ start/src/App.vue
@@ -1,3 +1,7 @@
 <script lang="ts" setup></script>
 
-<template></template>
+<template>
+  <f-skeleton animated :rows="8" round difference row-gap="20px" />
+
+  <!-- <f-skeleton round animated /> -->
+</template>